+++ src/index.ts
@@ -1,9 +1,5 @@
 export { default } from './components/StyledModal';
 
-export {
-  default as Portal,
-  collectPortals,
-  flushPortals
-} from './components/Portal';
+export { default as Portal, collectPortals, flushPortals } from './components/Portal';
 
 export { default as hasDom } from './utils/has-dom';