+++ src/styles/index.ts
@@ -1,7 +1,6 @@
 import { css } from 'styled-components';
-import * as Styled from 'styled-components/typings/styled-components';
 
-export const container: Styled.InterpolationValue[] = css`
+export const container = css`
   height: 100%;
   left: 0;
   overflow-x: hidden;
@@ -12,7 +11,7 @@ export const container: Styled.InterpolationValue[] = css`
   z-index: 9000;
 `;
 
-export const overscroll: Styled.InterpolationValue[] = css`
+export const overscroll = css`
   align-items: center;
   display: flex;
   justify-content: center;
@@ -21,6 +20,6 @@ export const overscroll: Styled.InterpolationValue[] = css`
   width: 100%;
 `;
 
-export const modal: Styled.InterpolationValue[] = css`
+export const modal = css`
   z-index: 9001;
 `;