+++ src/utils/has-dom.ts
@@ -1,3 +1,2 @@
 export default (): boolean =>
-  (typeof window as Partial<Window>) !== 'undefined' &&
-  (typeof window.document as Partial<Document>) !== 'undefined';
+  (typeof window as Partial<Window>) !== 'undefined' && (typeof window.document as Partial<Document>) !== 'undefined';