+++ stories/06-portal.tsx
@@ -7,10 +7,11 @@ storiesOf('styled-modal', module).add('Portal with custom target', () => {
   const target = 'error-display';
 
   return (
-    <React.Fragment>
+    <>
+      <h1>Portal can target any dom node by id</h1>
       <Portal target={target}>
         <p>This text is portaled to Storybook's default #error-display</p>
       </Portal>
-    </React.Fragment>
+    </>
   );
 });